Split the generic highlight core out of the audio highlighting (BL-16558)
Behavior-preserving refactor to prepare for the decodable/leveled reader
tools using the same ::highlight technique.

bookEdit/js/textHighlightManager.ts is now generic: it knows about the
CSS custom highlight registry and nothing about audio. An instance is
constructed with the set of highlight names it owns, and offers
canHighlight(), setHighlight(), clearHighlights(),
clearAllManagedHighlights() and hasDeadRanges(). The Highlight object's
priority is now settable through setHighlight(), which the reader tools
will need for nested (word inside sentence) highlights.

toolbox/talkingBook/audioHighlightManager.ts holds everything specific to
Talking Book: the highlight names, the current/split highlight rules, the
ui-enableHighlight and bloom-highlightSegment selection logic, and the
StyleEditor lookup of the user's chosen highlight colors. It delegates
the registry work to a TextHighlightManager.

The public methods audioRecording.ts calls keep their names, so its only
changes are the import, and renaming its field to audioHighlightManager.
The Talking Book tests pass unchanged.
