+++ src/BloomBrowserUI/bookEdit/js/textHighlightManager.ts
@@ -1,45 +1,30 @@
-import StyleEditor from "../StyleEditor/StyleEditor";
-
-const kSegmentClass = "bloom-highlightSegment";
-const kEnableHighlightClass = "ui-enableHighlight";
-const kDisableHighlightClass = "ui-disableHighlight";
-const kPostAudioSplitClass = "bloom-postAudioSplit";
-const kTextBoxRecordingMode = "textbox";
-
-const kCurrentHighlightBackgroundCssVar =
-    "--bloom-audio-current-highlight-background";
-const kCurrentHighlightColorCssVar = "--bloom-audio-current-highlight-color";
-
-// This manager translates Bloom's audio-highlight classes into the CSS highlight registry and
-// ::highlight pseudo-elements.
-// In rare cases, the browser can automatically move computed css into an inline style within a contenteditable, which
-// we suspect is causing BL-15300 where TBT highlighting gets stuck in the book. This method of highlighting without
-// modifying the dom should prevent that, and is also the direction we want to move in for highlighting.
-// The DOM still decides which pieces of text are eligible and marks them with the appropriate classes, but in the Edit
-// Tab the visible paint comes from ::highlight pseudo-elements instead of the element
-// background colors, which we continue to use in Bloom Player etc. - we will need to keep the original class and css
-// rules for a while so that old versions of Bloom player display the highlights, but a next step would be to make
-// newer versions of Bloom Player switch to using pseudo-elements to display highlights like we do here
-
-export const currentHighlightName = "bloom-audio-current";
-
-export const splitHighlightNames = [
-    "bloom-audio-split-1",
-    "bloom-audio-split-2",
-    "bloom-audio-split-3",
-] as const;
-
-const allManagedHighlightNames = [currentHighlightName, ...splitHighlightNames];
+// This manager paints text highlights using the CSS custom highlight registry
+// (CSS.highlights) and ::highlight() pseudo-elements, instead of by wrapping the text in
+// styled spans.
+//
+// In rare cases, the browser can automatically move computed css into an inline style within a
+// contenteditable, which we suspect is causing BL-15300 (Talking Book highlighting gets stuck in
+// the book) and BL-16558 (replacing a decodable-reader-marked word leaves a permanent
+// background-color span). Highlighting without modifying the dom prevents that, and is the
+// direction we want to move in for highlighting generally.
+//
+// The DOM still decides which pieces of text are eligible; only the visible paint comes from
+// ::highlight pseudo-elements. Where the highlight classes are also meaningful outside the Edit
+// Tab (as the Talking Book ones are, for Bloom Player), we continue to write those classes and
+// their background-color rules so older consumers still work.
+//
+// Each client (the Talking Book tool, the decodable/leveled reader tools) owns an instance
+// constructed with the set of highlight names it manages. Names are global to the document's
+// registry, so they must be unique across clients, and each needs a matching ::highlight(name)
+// rule in a stylesheet loaded into the document being highlighted.
 
 type HighlightRegistry = Map<string, unknown>;
-type HighlightConstructor = new (...ranges: Range[]) => unknown;
+// A real Highlight is setlike over its Ranges and has a settable numeric priority.
+type HighlightObject = Iterable<Range> & { priority?: number };
+type HighlightConstructor = new (...ranges: Range[]) => HighlightObject;
 
 function getDocumentWindow(contextNode: Node): Window | undefined {
     return contextNode.ownerDocument?.defaultView ?? undefined;
-}
-
-function getDocumentElement(contextNode: Node): HTMLElement | undefined {
-    return contextNode.ownerDocument?.documentElement ?? undefined;
 }
 
 function getHighlightRegistry(
@@ -67,57 +52,99 @@ function getHighlightConstructor(
     return docWindow?.Highlight;
 }
 
-// A StyleEditor instance used only to read the user's audio-highlight colors from a book's
-// userModifiedStyles sheet. We share StyleEditor's rule lookup rather than duplicating the
-// selector-matching logic, so the read and the write (StyleEditor.putAudioHiliteRulesInDom)
-// can never drift apart. supportFilesRoot is irrelevant for this read-only use.
-let styleEditorForColorLookup: StyleEditor | undefined;
-function getStyleEditorForColorLookup(): StyleEditor {
-    if (!styleEditorForColorLookup) {
-        styleEditorForColorLookup = new StyleEditor("/bloom/bookEdit");
+// Make a Range covering all of node's content, or undefined if there is nothing to highlight.
+export function makeRangeForNodeContents(node: Node): Range | undefined {
+    if (node.textContent === null || node.textContent.length === 0) {
+        return undefined;
     }
-    return styleEditorForColorLookup;
+
+    const ownerDocument = node.ownerDocument;
+    if (!ownerDocument) {
+        console.error(
+            "textHighlightManager.makeRangeForNodeContents() could not find ownerDocument for a highlighted node.",
+        );
+        return undefined;
+    }
+
+    const range = ownerDocument.createRange();
+    range.selectNodeContents(node);
+    return range;
 }
 
 export class TextHighlightManager {
-    // Remove all current and split highlights from the registry for the document containing contextNode.
+    // The highlight names this instance owns. clearAllManagedHighlights() removes exactly these.
+    private readonly managedHighlightNames: readonly string[];
+
+    public constructor(managedHighlightNames: readonly string[]) {
+        this.managedHighlightNames = managedHighlightNames;
+    }
+
+    // Returns true if the document containing contextNode supports the custom highlight API,
+    // so it is worth computing ranges at all. Test environments may not provide it.
+    public canHighlight(contextNode: Node): boolean {
+        return (
+            !!getHighlightRegistry(contextNode) &&
+            !!getHighlightConstructor(contextNode)
+        );
+    }
+
+    // Paint the given ranges under the given highlight name, replacing whatever that name was
+    // painting before. An empty list of ranges removes the highlight altogether.
+    // Where highlights can overlap, a higher priority wins; the default is 0.
+    public setHighlight(
+        name: string,
+        ranges: Range[],
+        contextNode: Node,
+        priority?: number,
+    ): void {
+        const registry = getHighlightRegistry(contextNode);
+        const Highlight = getHighlightConstructor(contextNode);
+        if (!registry || !Highlight) {
+            return;
+        }
+
+        if (ranges.length === 0) {
+            registry.delete(name);
+            return;
+        }
+
+        const highlight = new Highlight(...ranges);
+        if (priority !== undefined) {
+            highlight.priority = priority;
+        }
+        registry.set(name, highlight);
+    }
+
+    // Remove the named highlights from the registry for the document containing contextNode.
+    public clearHighlights(names: readonly string[], contextNode?: Node): void {
+        if (!contextNode) {
+            return;
+        }
+
+        const registry = getHighlightRegistry(contextNode);
+        if (!registry) {
+            return;
+        }
+
+        names.forEach((name) => registry.delete(name));
+    }
+
+    // Remove all of this instance's highlights from the registry for the document containing
+    // contextNode.
     public clearAllManagedHighlights(contextNode?: Node): void {
-        if (!contextNode) {
-            return;
-        }
-
-        const registry = getHighlightRegistry(contextNode);
-        if (!registry) {
-            return;
-        }
-
-        allManagedHighlightNames.forEach((name) => registry.delete(name));
+        this.clearHighlights(this.managedHighlightNames, contextNode);
     }
 
-    // Remove only the split (blue segment) highlights from the registry, leaving the current highlight intact.
-    public clearSplitHighlights(contextNode?: Node): void {
-        if (!contextNode) {
-            return;
-        }
-
-        const registry = getHighlightRegistry(contextNode);
-        if (!registry) {
-            return;
-        }
-
-        splitHighlightNames.forEach((name) => registry.delete(name));
-    }
-
-    // Returns true if the current-highlight entry in the registry for contextNode's document
-    // exists but any of its Ranges no longer cover live content, so the highlight is still
-    // registered but paints nothing (BL-15300). That happens two ways when page-setup code
-    // rewrites the DOM under the highlight:
+    // Returns true if the named entry in the registry for contextNode's document exists but any
+    // of its Ranges no longer cover live content, so the highlight is still registered but paints
+    // nothing (BL-15300). That happens two ways when page-setup code rewrites the DOM under the
+    // highlight:
     // - the range's node itself was replaced (e.g. CKEditor's initialization replaces the
     //   paragraph): the range points at a detached node;
     // - an ANCESTOR of the range's node was removed (e.g. re-appending the bloom-editables to
     //   reorder them): per the DOM spec the live Range is then COLLAPSED onto the
     //   still-connected former parent, so a connectedness check alone would miss it.
-    public currentHighlightHasDeadRanges(contextNode?: Node): boolean {
+    public hasDeadRanges(name: string, contextNode?: Node): boolean {
         if (!contextNode) {
             return false;
         }
@@ -127,9 +154,7 @@ export class TextHighlightManager {
         }
         // A real (Chromium) Highlight is setlike over its Ranges, so we can iterate it. Test
         // environments may register a non-iterable stand-in; treat those as healthy.
-        const highlight = registry.get(currentHighlightName) as
-            | Iterable<Range>
-            | undefined;
+        const highlight = registry.get(name) as HighlightObject | undefined;
         if (!highlight || typeof highlight[Symbol.iterator] !== "function") {
             return false;
         }
@@ -143,254 +168,5 @@ export class TextHighlightManager {
             }
         }
         return false;
-    }
-
-    // currentHighlight is the element currently selected for recording etc.
-    // It might be a span (sentence mode) or text box (text box mode).
-    // currentTextBox is either the same as currentHighlight (text box mode)
-    // or its TextBox ancestor (sentence mode).
-    // Adjust pseudo-element highlights to what they should be for this state of things.
-    public refreshHighlights(
-        currentHighlight: Element | null,
-        currentTextBox: HTMLElement | null,
-        suppressCurrentHighlight?: boolean,
-    ): void {
-        const contextNode = currentHighlight ?? currentTextBox;
-        if (!contextNode) {
-            return;
-        }
-
-        const registry = getHighlightRegistry(contextNode);
-        const Highlight = getHighlightConstructor(contextNode);
-        if (!registry || !Highlight) {
-            return;
-        }
-
-        if (suppressCurrentHighlight) {
-            allManagedHighlightNames.forEach((name) => registry.delete(name));
-            return;
-        }
-
-        // Split highlights (blue segments after a textbox recording) and the current highlight
-        // (yellow sentence) are mutually exclusive: split state replaces the yellow highlight.
-        if (this.shouldShowSplitHighlights(currentHighlight, currentTextBox)) {
-            registry.delete(currentHighlightName);
-            this.refreshSplitHighlights(currentTextBox, registry, Highlight);
-        } else {
-            splitHighlightNames.forEach((name) => registry.delete(name));
-            this.refreshCurrentHighlight(
-                currentHighlight,
-                currentTextBox,
-                registry,
-                Highlight,
-            );
-        }
-    }
-
-    private refreshCurrentHighlight(
-        currentHighlight: Element | null,
-        currentTextBox: HTMLElement | null,
-        registry: HighlightRegistry,
-        Highlight: HighlightConstructor,
-    ): void {
-        const highlightInfo = this.getCurrentHighlightInfo(
-            currentHighlight,
-            currentTextBox,
-        );
-        if (!highlightInfo || highlightInfo.ranges.length === 0) {
-            registry.delete(currentHighlightName);
-            return;
-        }
-
-        // enhance: don't check for highlight color settings changes so often
-        this.updateCurrentHighlightColors(highlightInfo.styleSource);
-        registry.set(
-            currentHighlightName,
-            new Highlight(...highlightInfo.ranges),
-        );
-    }
-
-    private refreshSplitHighlights(
-        currentTextBox: HTMLElement,
-        registry: HighlightRegistry,
-        Highlight: HighlightConstructor,
-    ): void {
-        // Cycle through 3 colors using a page-relative index so adjacent paragraphs
-        // never share the same color at their boundary.
-        const rangesByName = new Map<string, Range[]>();
-        splitHighlightNames.forEach((name) => rangesByName.set(name, []));
-
-        Array.from(
-            currentTextBox.querySelectorAll(`span.${kSegmentClass}`),
-        ).forEach((segment, index) => {
-            const highlightName =
-                splitHighlightNames[index % splitHighlightNames.length];
-            const ranges = rangesByName.get(highlightName);
-            ranges?.push(...this.getRangesForSegment(segment));
-        });
-
-        splitHighlightNames.forEach((name) => {
-            const ranges = rangesByName.get(name) ?? [];
-            if (ranges.length > 0) {
-                registry.set(name, new Highlight(...ranges));
-            } else {
-                registry.delete(name);
-            }
-        });
-    }
-
-    private getCurrentHighlightInfo(
-        currentHighlight: Element | null,
-        currentTextBox: HTMLElement | null,
-    ):
-        | {
-              ranges: Range[];
-              styleSource: Element;
-          }
-        | undefined {
-        if (!currentHighlight) {
-            return undefined;
-        }
-
-        // copilot says: fixHighlighting() can carve the visible pieces into nested ui-enableHighlight
-        // spans so punctuation or outer whitespace stays unpainted. Prefer those exact
-        // spans whenever they exist so the pseudo-highlight matches the background-color highlight behavior
-        const enabledDescendants = Array.from(
-            currentHighlight.querySelectorAll(`span.${kEnableHighlightClass}`),
-        );
-        const enabledRanges = enabledDescendants
-            .map((enabledSpan) => this.makeRange(enabledSpan))
-            .filter((range): range is Range => !!range);
-        if (enabledRanges.length > 0) {
-            return {
-                ranges: enabledRanges,
-                styleSource: enabledDescendants[0],
-            };
-        }
-
-        if (currentHighlight.classList.contains(kDisableHighlightClass)) {
-            return undefined;
-        }
-
-        if (currentHighlight === currentTextBox) {
-            const paragraphs = Array.from(currentTextBox.querySelectorAll("p"));
-            const paragraphRanges = paragraphs
-                .map((paragraph) => this.makeRange(paragraph))
-                .filter((range): range is Range => !!range);
-            if (paragraphRanges.length > 0) {
-                return {
-                    ranges: paragraphRanges,
-                    styleSource: paragraphs[0],
-                };
-            }
-        }
-
-        const wholeElementRange = this.makeRange(currentHighlight);
-        if (!wholeElementRange) {
-            return undefined;
-        }
-
-        return {
-            ranges: [wholeElementRange],
-            styleSource: currentHighlight,
-        };
-    }
-
-    // Set the CSS variables that control the ::highlight colors to match the user's chosen
-    // highlight color for the current text style, falling back to the default yellow.
-    private updateCurrentHighlightColors(styleSource: Element): void {
-        const documentElement = getDocumentElement(styleSource);
-        if (!documentElement) {
-            console.error(
-                "TextHighlightManager.updateCurrentHighlightColors() could not find documentElement for the style source.",
-            );
-            return;
-        }
-
-        const bloomEditable = styleSource.closest(".bloom-editable");
-        const styleName = bloomEditable
-            ? Array.from(bloomEditable.classList).find((c) =>
-                  c.endsWith("-style"),
-              )
-            : undefined;
-
-        // Read the user's chosen highlight colors for this style from the same
-        // userModifiedStyles rules that StyleEditor.putAudioHiliteRulesInDom writes, by
-        // delegating to StyleEditor's own lookup (looking in the page's document, not the
-        // toolbox's). When there is no such style, fall back to the default yellow/black.
-        const userColors = styleName
-            ? getStyleEditorForColorLookup().getAudioHiliteProps(
-                  styleName,
-                  styleSource.ownerDocument,
-              )
-            : undefined;
-
-        documentElement.style.setProperty(
-            kCurrentHighlightBackgroundCssVar,
-            userColors?.hiliteBgColor ?? "#febf00",
-        );
-        documentElement.style.setProperty(
-            kCurrentHighlightColorCssVar,
-            userColors?.hiliteTextColor ?? "black",
-        );
-    }
-
-    private shouldShowSplitHighlights(
-        currentHighlight: Element | null,
-        currentTextBox: HTMLElement | null,
-    ): currentTextBox is HTMLElement {
-        if (!currentHighlight || !currentTextBox) {
-            return false;
-        }
-
-        if (currentHighlight !== currentTextBox) {
-            return false;
-        }
-
-        if (currentTextBox.classList.contains(kDisableHighlightClass)) {
-            return false;
-        }
-
-        // Split highlights are only for textbox recordings after AudioRecording has
-        // split the textbox into segment spans and marked it as post-split.
-        return (
-            currentTextBox.classList.contains(kPostAudioSplitClass) &&
-            currentTextBox
-                .getAttribute("data-audiorecordingmode")
-                ?.toLowerCase() === kTextBoxRecordingMode
-        );
-    }
-
-    private getRangesForSegment(segment: Element): Range[] {
-        const enabledRanges = Array.from(
-            segment.querySelectorAll(`span.${kEnableHighlightClass}`),
-        )
-            .map((enabledSpan) => this.makeRange(enabledSpan))
-            .filter((range): range is Range => !!range);
-
-        if (enabledRanges.length > 0) {
-            return enabledRanges;
-        }
-
-        const wholeSegmentRange = this.makeRange(segment);
-        return wholeSegmentRange ? [wholeSegmentRange] : [];
-    }
-
-    private makeRange(node: Node): Range | undefined {
-        if (node.textContent === null || node.textContent.length === 0) {
-            return undefined;
-        }
-
-        const ownerDocument = node.ownerDocument;
-        if (!ownerDocument) {
-            console.error(
-                "TextHighlightManager.makeRange() could not find ownerDocument for a highlighted node.",
-            );
-            return undefined;
-        }
-
-        const range = ownerDocument.createRange();
-        range.selectNodeContents(node);
-        return range;
     }
 }
